pyproject.toml
[project]
name = "pumc-course-select"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numexpr>=2.14.1",
    "oldest-supported-numpy>=2023.12.21",
    "opencv-python>=4.11.0.86",
    "pillow>=12.1.0",
    "pytesseract>=0.3.13",
    "requests>=2.32.5",
]
